+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- [24-12-8] Fix: Weekly RTTY mults. Add RTC to Weekly RTTY.
 - [24-12-6] Add RTC to K1USN.
 -[24-12-5-1] ARRL 160 gets rtc.
 - [24-12-5] Add 'real time' score posting to external sites.